+++ website/docs/_blogs/2024-12-02-ReasoningAgent2/index.mdx
@@ -81,11 +81,12 @@ import os
 from autogen import (
     AssistantAgent,
     UserProxyAgent,
+)
+from autogen.agents.experimental.reasoning_agent import (
     ReasoningAgent,
     ThinkNode,
     visualize_tree
 )
-
 
 # Configure the model
 config_list = [{"api_type": "openai", "model": "gpt-4", "api_key": os.environ.get("OPENAI_API_KEY")}]
@@ -217,7 +218,7 @@ new_node = pickle.load(open("reasoning_tree.pkl", "rb"))
 This step finds the best trajectory in the thought tree and converts it to a SFT dataset as a sequence of strings. The best trajectory is determined by following the highest-scoring path from root to leaf.
 
 ```python
-from autogen.agentchat.contrib.reasoning_agent import extract_sft_dataset
+from autogen.agents.experimental.reasoning_agent import extract_sft_dataset
 
 # Example usage
 sft_data = extract_sft_dataset(reason_agent._root)
@@ -228,7 +229,7 @@ json.dump(sft_data, open("sft_data.json", "w"), indent=2)
 This step generates preference pairs by comparing sibling nodes in the tree. For each parent node with multiple children, we create training pairs where the higher-scored response is marked as preferred over the lower-scored one.
 
 ```python
-from autogen.agentchat.contrib.reasoning_agent import extract_rlhf_preference_dataset
+from autogen.agents.experimental.reasoning_agent import extract_rlhf_preference_dataset
 
 # Example usage
 rlhf_data = extract_rlhf_preference_dataset(reason_agent._root)
