+++ website/docs/user-guide/reference-agents/reasoningagent.mdx
@@ -1035,7 +1035,7 @@ with open("reasoning_tree.json", "w") as f:
 # recover the node
 new_node = ThinkNode.from_dict(json.load(open("reasoning_tree.json")))  # noqa: SIM115
 
-from autogen.agentchat.contrib.reasoning_agent import extract_rlhf_preference_dataset, extract_sft_dataset
+from autogen.agents.experimental.reasoning_agent import extract_rlhf_preference_dataset, extract_sft_dataset
 
 sft_data = extract_sft_dataset(reason_agent._root)
 rlhf_data = extract_rlhf_preference_dataset(reason_agent._root)
